Shared-Engine Parity: Refactor Model Maturity Assessment into engine package
Delegates the computation of model maturity score, level, parameter breakdown,
factors, and recommendations from packages/mcp-server handler to the shared
engine function assessModelMaturity in @bngplayground/engine. This removes duplication
and enforces strict shared-engine parity.
